Работает /user/register в package Userdetails который используется для Security объявлен метод getUsername, поэтому связка для входа username|password, а не login|password.
/user/login
/user/history/{login}
/room/{room_uuid}
/room/booking/{room_uuid} (get)
/room/booking/{room_uuid} (post)

В процессе
/room/booking/{room_uuid} (delete)

+++ src/main/java/NC/mtroom/room/impl/entity/Equipment.java
@@ -17,7 +17,7 @@ public class Equipment {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer EqID;
+    private Long EqID;
 
     private String name;
